pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "isodate>=0.7.2",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
